Basic logic and routing is rewritten for backend version

+++ src/main/java/com/kiselev/time/model/response/CalculatorResponse.java
@@ -13,5 +13,5 @@ public class CalculatorResponse {
 
     private Long cleanTime;
 
-    private Long percents;
+    private Double percents;
 }
